requirements.txt
feedparser==6.0.11
jinja2==3.1.4
requests==2.32.5
python-dateutil==2.9.0.post0
markdown==3.7
